// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Size = {
  XS: 'XS',
  S: 'S',
  M: 'M',
  L: 'L',
  XL: 'XL',
  XXL: 'XXL',
  XXXL: 'XXXL'
} as const

export type Size = (typeof Size)[keyof typeof Size]


export const Gender = {
  men: 'men',
  women: 'women',
  kid: 'kid',
  unisex: 'unisex',
  other: 'other'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const Role = {
  admin: 'admin',
  user: 'user'
} as const

export type Role = (typeof Role)[keyof typeof Role]